Important update to the  Libero project. Working UART
Due to issues where the project could not be modified without causing
the project to fail, the Libero project was re-created from scratch.

More specifically, the project cloned from github would work fine, but
after only changing the pin assignment of the uart pins to be correct,
the debugger can't attach to upload software, and trying to read flash
with the libero debugger gives bus errors. I tried uploading updated
software before updating the fpga, changing to different pins, and
changing the CAN pins, but nothing would result in a workable project.

There are some slight changes but nothing that should have an impact on
the current or future software.

The new project has the correct pinout for the breadboard model with
mram,external flash,dev board flash (not-tested), can, uart, and some of
the LEDs.

The CoreSPI peripherals were renamed in the Libero project, so the SPI
driver was changed to use the new name.

This project has been tested, and the UART task, Filesystem test, CAN
RX/TX tasks, LED task, and mram task all worked.